add ajustes exception

+++ src/main/java/com/projetoentrevista/services/PaisService.java
@@ -4,6 +4,7 @@ import com.projetoentrevista.dto.DadosListagemPaisDTO;
 import com.projetoentrevista.dto.PaisDTO;
 import com.projetoentrevista.entities.Pais;
 import com.projetoentrevista.exceptions.ResourceNotFoundException;
+import com.projetoentrevista.exceptions.ResourceNotValidException;
 import com.projetoentrevista.repositories.PaisRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,7 +21,7 @@ public class PaisService {
 
     public PaisDTO listarPais(String codigoInternacinalIso) {
         Optional<Pais> optionalPais = repository.findByCodigoInternacionalIso(codigoInternacinalIso.toUpperCase());
-        return optionalPais.orElseThrow(() -> new ResourceNotFoundException("Código Internacional Iso " + codigoInternacinalIso + " não foi encontrado.")).toPaisDTO();
+        return optionalPais.orElseThrow(() -> new ResourceNotFoundException("Código Internacional Iso " + codigoInternacinalIso + " não foi encontrado na base de dados.")).toPaisDTO();
     }
 
     public List<DadosListagemPaisDTO> listagemPaises() {
@@ -29,16 +30,19 @@ public class PaisService {
     }
 
     public PaisDTO adicionarPais(Pais pais) {
+        if (repository.existsByCodigoInternacionalIso(pais.getCodigoInternacionalIso())) {
+            throw new ResourceNotValidException("País com código internacional Iso " + pais.getCodigoInternacionalIso() + " já está cadastrado na base de dados");
+        }
         Pais paisRetornado = repository.save(pais);
         PaisDTO paisDTO = new PaisDTO(paisRetornado);
         return paisDTO;
     }
 
     public void deletarpais(Long id) {
-        try {
-            repository.deleteById(id);
-        } catch (Exception ex) {
-            throw new ResourceNotFoundException("O ID " + id + " não foi encontrado.");
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("O ID " + id + " não foi encontrado na base de dados.");
         }
+        repository.deleteById(id);
     }
+
 }
